requirements.txt
py==1.4.31
pytest==3.0.3
six==1.10.0
testinfra==1.4.2
requests==2.11.1
docker-py==1.10.6
